requirements.txt
Django==5.2.1
djangorestframework==3.16.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
Pillow==10.3.0